Refactor: 새 아키텍처에 맞는 디렉토리로 변경
