+++ src/main/java/com/moyo/backend/domain/pr_review/data_access/PrReviewRepositoryImpl.java
@@ -1,4 +1,4 @@
-package com.moyo.backend.domain.pr_review.infrastructure;
+package com.moyo.backend.domain.pr_review.data_access;
 
 import java.util.Optional;
 
@@ -8,9 +8,8 @@ import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Repository;
 
-import com.moyo.backend.domain.pr_review.application.port.PrReviewRepository;
-import com.moyo.backend.domain.pr_review.domain.PrReview;
-import com.moyo.backend.domain.pr_review.domain.position.Position;
+import com.moyo.backend.domain.pr_review.implement.PrReview;
+import com.moyo.backend.domain.pr_review.implement.Position;
 import com.moyo.backend.domain.user.implement.User;
 
 @Repository
