+++ src/main/java/com/moyo/backend/domain/pr_review/presentation/PrReviewController.java
@@ -17,7 +17,7 @@ import org.springframework.web.bind.annotation.RestController;
 
 import com.moyo.backend.common.response.ApiResponse;
 import com.moyo.backend.domain.auth.oauth.dto.GithubOAuth2User;
-import com.moyo.backend.domain.pr_review.application.PrReviewService;
+import com.moyo.backend.domain.pr_review.business.PrReviewService;
 import com.moyo.backend.domain.pr_review.dto.request.PrReviewCreateRequestDto;
 import com.moyo.backend.domain.pr_review.dto.request.PrReviewListRequestDto;
 import com.moyo.backend.domain.pr_review.dto.request.PrReviewUpdateRequestDto;
